Make mock constants folder
