+++ src/common/index.ts
@@ -7,3 +7,4 @@ export * from "./urlUtils";
 export * from "./fileUtils";
 export * from "./types";
 export * from "./urlUtils";
+export * from "./mockConstants";
