+++ src/common/mockConstants/mockFile.tsx
@@ -1,0 +1,12 @@
+export const mockFile = {
+  name: "hello-world.txt",
+  size: 1000,
+  type: "",
+  webkitRelativePath: "/",
+  lastModified: 1234567890,
+  preview: "",
+  arrayBuffer: (() => {}) as any, // eslint-disable-line
+  slice: (() => {}) as any, // eslint-disable-line
+  stream: (() => {}) as any, // eslint-disable-line
+  text: (() => {}) as any, // eslint-disable-line
+};
