+++ src/common/mockConstants/mockProfile.tsx
@@ -1,0 +1,14 @@
+import { Profile } from "modules/session";
+
+export const mockProfile: Profile = {
+  id: "ABCD1234",
+  oauthId: "",
+  oauthType: "",
+  email: "example@email.com",
+  firstName: "John",
+  lastName: "Smith",
+  nickname: "Johnny",
+  pictureUrl: "",
+  role: "Producer",
+  genre: "Instrumental",
+};
