+++ src/common/mockConstants/mockSession.tsx
@@ -1,0 +1,8 @@
+import { SessionState } from "modules/session";
+import { mockProfile } from "./mockProfile";
+
+export const mockSession: SessionState = {
+  isLoggedIn: true,
+  errorMessage: "",
+  profile: mockProfile,
+};
