+++ src/common/mockConstants/mockUseGetSongsQueryResponse.ts
@@ -1,0 +1,181 @@
+export const mockUseGetSongsQueryResponse = {
+  status: "fulfilled",
+  endpointName: "getSongs",
+  requestId: "QWP7izjAX1LzoxdHlTAhB",
+  startedTimeStamp: 1655242844541,
+  data: [
+    {
+      id: "c3bc680b-db0a-4e7e-b008-044e53aa02c2",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:59.707476",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "b3a79fc4-a979-4525-972f-dc4d92e69d64",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:37:00.350493",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "7ba956dd-0cb8-488d-a5b8-6f7b56c15a95",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:37:01.574709",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "131f1899-5f97-468d-a069-ab2fec4a5122",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:39:24.238835",
+      title: "My name is Jonas",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+  ],
+  fulfilledTimeStamp: 1655242844776,
+  isUninitialized: false,
+  isLoading: false,
+  isSuccess: true,
+  isError: false,
+  currentData: [
+    {
+      id: "cdd3ef09-9f26-4f76-ab74-d858ac44eaa1",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:56.629345",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "b3fb8447-b7bf-41fe-b9ff-970265e0f718",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:57.204444",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "a4626221-3c29-47dc-9651-d2adccec227a",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:58.949314",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "d0b4c8ae-3ad8-43c5-924a-5dd8ebef7dce",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:59.186792",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "42281de5-8146-4e17-b216-40ed002dab61",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:59.445258",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "c3bc680b-db0a-4e7e-b008-044e53aa02c2",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:36:59.707476",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "b3a79fc4-a979-4525-972f-dc4d92e69d64",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:37:00.350493",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "7ba956dd-0cb8-488d-a5b8-6f7b56c15a95",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:37:01.574709",
+      title: "Test",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "131f1899-5f97-468d-a069-ab2fec4a5122",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:39:24.238835",
+      title: "My name is Jonas",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "2d78237a-33ad-4ec1-b0ff-5fca2940a129",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:40:00.465598",
+      title: "Say It Ain't So",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+    {
+      id: "5a927c02-7b2d-483b-b684-115b5541edaf",
+      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
+      createdAt: "2022-06-12T18:40:11.994142",
+      title: "Karma Police",
+      genre: "Rock",
+      coverArtUrl:
+        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
+      description: "Test description",
+      credits: "Produced by Max Martin and Oscar Holter",
+    },
+  ],
+  isFetching: false,
+};
