--- src/pages/home/library/test/mockUseGetSongsQueryResponse.ts
@@ -1,610 +1,0 @@
-export const mockUseGetSongsQueryResponse = {
-  status: "fulfilled",
-  endpointName: "getSongs",
-  requestId: "QWP7izjAX1LzoxdHlTAhB",
-  startedTimeStamp: 1655242844541,
-  data: [
-    {
-      id: "fd620a2e-8de7-43d5-ab36-0c8589df7047",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:14:32.259011",
-      title: "Test",
-      genre: "Alt Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/1/14/Inrainbowscover.png",
-      description: "Song by Radiohead",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a3d500cf-ddf1-45a8-9d19-2fb0845ae573",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:31:31.403317",
-      title: "Test",
-      genre: "Hip Hop",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/d/d3/JColeKOD.jpg",
-      description: "Song by J Cole",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b933329c-2899-4a9d-a62d-56ab68d38db0",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:32:49.015830",
-      title: "Test",
-      genre: "Hip Hop",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Graduation_%28album%29.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "47692a10-705b-4be3-a415-13edc0b9f1b9",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:33:21.879313",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/d/df/RedHotChiliPeppersCalifornication.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a57869c2-de62-4951-88f5-3e5742ebce3b",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:35:15.182831",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/3/3f/LogicEverybody.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "34a0d635-6697-4ce4-aff8-50b3ba0f3287",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:30.946745",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b51b3783-e12d-42e5-817c-8028c53f9619",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:46.100759",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "8c556988-e697-4135-b086-8d3627d77fe1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:46.757665",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "33ec7e9d-f98b-4f27-8380-99ba5055e1b1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:47.533086",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "6f07d34a-b85b-4f2b-b66b-0da30064f9bb",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:48.116134",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "3e0666d3-37e5-4921-b9e8-eae23ee7ac48",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:48.826921",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "d93dad4f-bc55-47d7-98be-b7fbf1b254b1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:54.887695",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "51f0cd36-842f-42df-8df5-c2ba37a0cee2",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:55.252776",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "c36940c9-4336-4728-9275-e322c1839374",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:55.748676",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "0ab2b9dc-b979-4d39-b75d-b58b00ad5787",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.058042",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "2424ff73-3a01-4543-a426-2f2024b4e56e",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.342510",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "cdd3ef09-9f26-4f76-ab74-d858ac44eaa1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.629345",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b3fb8447-b7bf-41fe-b9ff-970265e0f718",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:57.204444",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a4626221-3c29-47dc-9651-d2adccec227a",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:58.949314",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "d0b4c8ae-3ad8-43c5-924a-5dd8ebef7dce",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.186792",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "42281de5-8146-4e17-b216-40ed002dab61",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.445258",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "c3bc680b-db0a-4e7e-b008-044e53aa02c2",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.707476",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b3a79fc4-a979-4525-972f-dc4d92e69d64",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:37:00.350493",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "7ba956dd-0cb8-488d-a5b8-6f7b56c15a95",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:37:01.574709",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "131f1899-5f97-468d-a069-ab2fec4a5122",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:39:24.238835",
-      title: "My name is Jonas",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "2d78237a-33ad-4ec1-b0ff-5fca2940a129",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:40:00.465598",
-      title: "Say It Ain't So",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "5a927c02-7b2d-483b-b684-115b5541edaf",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:40:11.994142",
-      title: "Karma Police",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-  ],
-  fulfilledTimeStamp: 1655242844776,
-  isUninitialized: false,
-  isLoading: false,
-  isSuccess: true,
-  isError: false,
-  currentData: [
-    {
-      id: "fd620a2e-8de7-43d5-ab36-0c8589df7047",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:14:32.259011",
-      title: "Test",
-      genre: "Alt Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/1/14/Inrainbowscover.png",
-      description: "Song by Radiohead",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a3d500cf-ddf1-45a8-9d19-2fb0845ae573",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:31:31.403317",
-      title: "Test",
-      genre: "Hip Hop",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/d/d3/JColeKOD.jpg",
-      description: "Song by J Cole",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b933329c-2899-4a9d-a62d-56ab68d38db0",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:32:49.015830",
-      title: "Test",
-      genre: "Hip Hop",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Graduation_%28album%29.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "47692a10-705b-4be3-a415-13edc0b9f1b9",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:33:21.879313",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/d/df/RedHotChiliPeppersCalifornication.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a57869c2-de62-4951-88f5-3e5742ebce3b",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:35:15.182831",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/3/3f/LogicEverybody.jpg",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "34a0d635-6697-4ce4-aff8-50b3ba0f3287",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:30.946745",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b51b3783-e12d-42e5-817c-8028c53f9619",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:46.100759",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "8c556988-e697-4135-b086-8d3627d77fe1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:46.757665",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "33ec7e9d-f98b-4f27-8380-99ba5055e1b1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:47.533086",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "6f07d34a-b85b-4f2b-b66b-0da30064f9bb",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:48.116134",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "3e0666d3-37e5-4921-b9e8-eae23ee7ac48",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:48.826921",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "d93dad4f-bc55-47d7-98be-b7fbf1b254b1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:54.887695",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "51f0cd36-842f-42df-8df5-c2ba37a0cee2",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:55.252776",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "c36940c9-4336-4728-9275-e322c1839374",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:55.748676",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "0ab2b9dc-b979-4d39-b75d-b58b00ad5787",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.058042",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "2424ff73-3a01-4543-a426-2f2024b4e56e",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.342510",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "cdd3ef09-9f26-4f76-ab74-d858ac44eaa1",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:56.629345",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b3fb8447-b7bf-41fe-b9ff-970265e0f718",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:57.204444",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "a4626221-3c29-47dc-9651-d2adccec227a",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:58.949314",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "d0b4c8ae-3ad8-43c5-924a-5dd8ebef7dce",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.186792",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "42281de5-8146-4e17-b216-40ed002dab61",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.445258",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "c3bc680b-db0a-4e7e-b008-044e53aa02c2",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:36:59.707476",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "b3a79fc4-a979-4525-972f-dc4d92e69d64",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:37:00.350493",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "7ba956dd-0cb8-488d-a5b8-6f7b56c15a95",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:37:01.574709",
-      title: "Test",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "131f1899-5f97-468d-a069-ab2fec4a5122",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:39:24.238835",
-      title: "My name is Jonas",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "2d78237a-33ad-4ec1-b0ff-5fca2940a129",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:40:00.465598",
-      title: "Say It Ain't So",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-    {
-      id: "5a927c02-7b2d-483b-b684-115b5541edaf",
-      ownerId: "29557f40-8866-4215-a739-1b2c853aa672",
-      createdAt: "2022-06-12T18:40:11.994142",
-      title: "Karma Police",
-      genre: "Rock",
-      coverArtUrl:
-        "https://upload.wikimedia.org/wikipedia/en/7/70/Weezer_-_Blue_Album.png",
-      description: "Test description",
-      credits: "Produced by Max Martin and Oscar Holter",
-    },
-  ],
-  isFetching: false,
-};
